requirements.txt
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
groq>=0.9.0
pypdf>=4.2.0
streamlit>=1.35.0
numpy>=1.26.0
